HoverComponent editability, minor TankModel edit
This is basically me just procrastinating the in-air movement...

Change-log:
HoverComponent editability - It is now possible to edit the values of HoverComponent through PlayerCar in "PlayerPhysics".
TankModel - Tweaked the Tank model to automatically point forwards, also some size tweaking.
Camera - Fixed Camera so that it points in front of the player instead of diagonally to the left.
Energy - Fixed so that the log actually shows the Energy number upon picking up Energy.
& other changes I cannot remember as of now.